+++ examples/modal-route-with-outlet/src/images.ts
@@ -1,0 +1,28 @@
+let IMAGES = [
+  {
+    id: 0,
+    title: "Enjoying a cup of coffee",
+    src: "https://images.unsplash.com/photo-1631016800696-5ea8801b3c2a?crop=entropy&cs=tinysrgb&fit=crop&fm=jpg&h=400&ixid=MnwxfDB8MXxyYW5kb218MHx8fHx8fHx8MTYzMzM2Mzg4Ng&ixlib=rb-1.2.1&q=80&w=400",
+  },
+  {
+    id: 1,
+    title: "Magical winter sunrise",
+    src: "https://images.unsplash.com/photo-1618824834718-92f8469a4dd1?crop=entropy&cs=tinysrgb&fit=crop&fm=jpg&h=400&ixid=MnwxfDB8MXxyYW5kb218MHx8fHx8fHx8MTYzMzM2NDAzMw&ixlib=rb-1.2.1&q=80&w=400",
+  },
+  {
+    id: 2,
+    title: "Dalmatian and pumpkins",
+    src: "https://images.unsplash.com/photo-1633289944756-6295be214e16?crop=entropy&cs=tinysrgb&fit=crop&fm=jpg&h=400&ixid=MnwxfDB8MXxyYW5kb218MHx8fHx8fHx8MTYzMzM2NDA3Nw&ixlib=rb-1.2.1&q=80&w=400",
+  },
+  {
+    id: 3,
+    title: "Fall into Autumn 🍂🐶",
+    src: "https://images.unsplash.com/photo-1633172905740-2eb6730c95b4?crop=entropy&cs=tinysrgb&fit=crop&fm=jpg&h=400&ixid=MnwxfDB8MXxyYW5kb218MHx8fHx8fHx8MTYzMzM2NDEwMg&ixlib=rb-1.2.1&q=80&w=400",
+  },
+];
+
+function getImageById(id: number) {
+  return IMAGES.find((image) => image.id === id);
+}
+
+export { IMAGES, getImageById };